Replace datatype for time value from last commit
In last commit, the builtin_interfaces::msg::Time datatype was
suggested to be used for everything. However, there exist some
limitations in how that message type is supposed to be utilized
in the ROS framework. So, here a single int64_t value is used
to represent all time values in the units of nanoseconds. This
solves all the problems associated with time representation.