Remove the checks for lockstep
This was a bad way to implement lockstep. With this method, everything
gets into a lock, mainly because there is no common clock source that
the complete ROS system can rely on right now. A better way to create
a common clock source would be to publish simulation time values over
a '/clock' topic. Then run all the nodes with use_sim_time argument.
This will cause the ROS to use '/clock' topic clock values instead of
the time values provided by the operating system.